+++ com.unity.cinemachine/Editor/Experimental/CmCameraEditor.cs
@@ -63,9 +63,9 @@ namespace Cinemachine.Editor
             var ux = new VisualElement();
 
             m_CameraUtility.AddCameraStatus(ux);
-            ux.Add(new PropertyField(serializedTarget.FindProperty(() => Target.m_Priority)));
             ux.Add(new PropertyField(serializedTarget.FindProperty(() => Target.m_StandbyUpdate)));
             ux.Add(new PropertyField(serializedTarget.FindProperty(() => Target.Transitions)));
+            ux.Add(new PropertyField(serializedTarget.FindProperty(() => Target.CameraPriority)));
             
             ux.AddHeader("Camera");
             var lensProperty = serializedTarget.FindProperty(() => Target.Lens);