+++ com.unity.cinemachine/Runtime/Behaviours/CinemachineFreeLook.cs
@@ -132,12 +132,10 @@ namespace Cinemachine
         private float m_LegacyHeadingBias = float.MaxValue;
         bool mUseLegacyRigDefinitions = false;
 
-        /// <summary>Enforce bounds for fields, when changed in inspector.</summary>
-        protected override void OnValidate()
+        protected override void LegacyUpgrade(int streamedVersion)
         {
-            base.OnValidate();
+            base.LegacyUpgrade(streamedVersion);
 
-            // Upgrade after a legacy deserialize
             if (m_LegacyHeadingBias != float.MaxValue)
             {
                 m_Heading.m_Bias= m_LegacyHeadingBias;
@@ -152,6 +150,11 @@ namespace Cinemachine
                 m_Transitions.m_BlendHint = m_LegacyBlendHint;
                 m_LegacyBlendHint = BlendHint.None;
             }
+        }
+        
+        /// <summary>Enforce bounds for fields, when changed in inspector.</summary>
+        void OnValidate()
+        {
             m_YAxis.Validate();
             m_XAxis.Validate();
             m_RecenterToTargetHeading.Validate();
@@ -635,8 +638,8 @@ namespace Cinemachine
                 if (rig == null)
                     continue;
                 rig.m_ExcludedPropertiesInInspector = m_CommonLens
-                    ? new string[] { "m_Script", "Header", "Extensions", "m_Priority", "m_Transitions", "m_Follow", "m_StandbyUpdate", "m_Lens" }
-                    : new string[] { "m_Script", "Header", "Extensions", "m_Priority", "m_Transitions", "m_Follow", "m_StandbyUpdate" };
+                    ? new string[] { "m_Script", "Header", "Extensions", "CameraPriority", "m_Transitions", "m_Follow", "m_StandbyUpdate", "m_Lens" }
+                    : new string[] { "m_Script", "Header", "Extensions", "CameraPriority", "m_Transitions", "m_Follow", "m_StandbyUpdate" };
                 rig.m_LockStageInInspector = new CinemachineCore.Stage[] { CinemachineCore.Stage.PositionControl };
             }
 #endif