+++ com.unity.cinemachine/Runtime/Core/CinemachineVirtualCameraBase.cs
@@ -28,36 +28,25 @@ namespace Cinemachine
     /// Unity cameras simultaneously.
     /// </summary>
     [SaveDuringPlay]
-    public abstract class CinemachineVirtualCameraBase : MonoBehaviour, ICinemachineCamera
+    public abstract class CinemachineVirtualCameraBase : MonoBehaviour, ICinemachineCamera, ISerializationCallbackReceiver
     {
         // TODO: remove m_ExcludedPropertiesInInspector
         /// <summary>Inspector control - Use for hiding sections of the Inspector UI.</summary>
         [HideInInspector, SerializeField, NoSaveDuringPlay]
         public string[] m_ExcludedPropertiesInInspector = new string[] { "m_Script" };
 
+        // TODO: remove m_LockStageInInspector
         /// <summary>Inspector control - Use for enabling sections of the Inspector UI.</summary>
         [HideInInspector, SerializeField, NoSaveDuringPlay]
         public CinemachineCore.Stage[] m_LockStageInInspector;
-
-        /// <summary>Version that was last streamed, for upgrading legacy</summary>
-        public int ValidatingStreamVersion
-        {
-            get { return m_OnValidateCalled ? m_ValidatingStreamVersion : CinemachineCore.kStreamingVersion; }
-            private set { m_ValidatingStreamVersion = value; }
-        }
-        private int m_ValidatingStreamVersion = 0;
-        private bool m_OnValidateCalled = false;
-
-        [HideInInspector, SerializeField, NoSaveDuringPlay]
-        private int m_StreamingVersion;
 
         /// <summary>The priority will determine which camera becomes active based on the
         /// state of other cameras and this camera.  Higher numbers have greater priority.
         /// </summary>
         [NoSaveDuringPlay]
         [Tooltip("The priority will determine which camera becomes active based on the state of "
-            + "other cameras and this camera.  Higher numbers have greater priority.")]
-        public int m_Priority = 10;
+            + "other cameras and this camera.  Higher numbers have greater priority. 0 is default.")]
+        public CameraPriority CameraPriority;
 
         /// <summary>A sequence number that represents object activation order of vcams.  
         /// Used for priority sorting.</summary>
@@ -105,6 +94,38 @@ namespace Cinemachine
             + "Set this to tune for performance. Most of the time Never is fine, "
             + "unless the virtual camera is doing shot evaluation.")]
         public StandbyUpdateMode m_StandbyUpdate = StandbyUpdateMode.RoundRobin;
+
+        //============================================================================
+        // Legacy streaming support
+
+        [HideInInspector, SerializeField, NoSaveDuringPlay]
+        private int m_StreamingVersion;
+
+        [HideInInspector, SerializeField, FormerlySerializedAs("m_Priority")]
+        int m_LegacyPriority = 10;
+
+        void ISerializationCallbackReceiver.OnAfterDeserialize()
+        {
+            if (m_StreamingVersion != CinemachineCore.kStreamingVersion)
+            {
+                LegacyUpgrade(m_StreamingVersion);
+                m_StreamingVersion = CinemachineCore.kStreamingVersion;
+            }
+        }
+
+        void ISerializationCallbackReceiver.OnBeforeSerialize() {}
+
+        /// <summary>
+        /// Override this to handle any upgrades necessitated by a streaming version change
+        /// </summary>
+        /// <param name="streamedVersion">The version that was streamed</param>
+        protected virtual void LegacyUpgrade(int streamedVersion)
+        {
+            if (streamedVersion < 20220601)
+                CameraPriority = new CameraPriority { Priority = m_LegacyPriority, UseCustomPriority = true };
+        }
+
+        //============================================================================
 
         /// <summary>
         /// Query components and extensions for the maximum damping time.
@@ -355,9 +376,10 @@ namespace Cinemachine
 
         /// <summary>Get the Priority of the virtual camera.  This determines its placement
         /// in the CinemachineCore's queue of eligible shots.</summary>
-        public int Priority { 
-            get => m_Priority;
-            set => m_Priority = value;
+        public int Priority 
+        { 
+            get => CameraPriority.UseCustomPriority ? CameraPriority.Priority : 0;
+            set => CameraPriority = new CameraPriority { Priority = value, UseCustomPriority = true };
         }
 
         /// <summary>Hint for blending to and from this virtual camera</summary>
@@ -567,16 +589,6 @@ namespace Cinemachine
             return null;
         }
 
-        /// <summary>Enforce bounds for fields, when changed in inspector.
-        /// Call base class implementation at the beginning of overridden method.
-        /// After base method is called, ValidatingStreamVersion will be valid.</summary>
-        protected virtual void OnValidate()
-        {
-            m_OnValidateCalled = true;
-            ValidatingStreamVersion = m_StreamingVersion;
-            m_StreamingVersion = CinemachineCore.kStreamingVersion;
-        }
-
         /// <summary>Base class implementation adds the virtual camera from the priority queue.</summary>
         protected virtual void OnEnable()
         {
@@ -610,7 +622,7 @@ namespace Cinemachine
         /// <summary>Base class implementation makes sure the priority queue remains up-to-date.</summary>
         protected virtual void Update()
         {
-            if (m_Priority != m_QueuePriority)
+            if (Priority != m_QueuePriority)
             {
                 UpdateVcamPoolStatus(); // Force a re-sort
             }
@@ -664,7 +676,7 @@ namespace Cinemachine
             CinemachineCore.Instance.RemoveActiveCamera(this);
             if (m_parentVcam == null && isActiveAndEnabled)
                 CinemachineCore.Instance.AddActiveCamera(this);
-            m_QueuePriority = m_Priority;
+            m_QueuePriority = Priority;
         }
 
         /// <summary>When multiple virtual cameras have the highest priority, there is