+++ com.unity.cinemachine/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - VirtualCamera refactor: CinemachineVirtualCamera is now CmCamera.
 - FreeLook refactor: CinemachineFreeLook is now CmCamera with FreeLook Modifier.
 - Combine Follow and LookAt Targets to single Tracking Target with optional LookAt Target.
+- Add flag for custom priority setting.
 
 
 ## UNRELEASED